redet_py36_cu110_requirements.txt
absl-py==1.0.0
addict==2.4.0
albumentations @ file:///home/yyw/yyf/projects/albumentations.zip
bleach==1.5.0
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.7
cycler==0.10.0
Cython @ file:///tmp/build/80754af9/cython_1626256602391/work
dataclasses @ file:///tmp/build/80754af9/dataclasses_1614363715916/work
decorator==4.4.0
distro==1.6.0
e2cnn==0.2.1
e2cnn-py36 @ file:///home/yyw/yyf/projects/swin-redet-test/e2cnn-legacy_py3.6.zip
enum34==1.1.6
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
html5lib==0.9999999
idna==3.3
imageio==2.5.0
importlib-metadata==4.8.2
joblib==1.1.0
kiwisolver==1.0.1
Markdown==3.1
matplotlib==3.0.3
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mmcv-full==1.3.16
mmpycocotools==12.0.3
mpmath==1.2.1
msgpack==0.6.1
msgpack-numpy==0.4.4.2
natsort==6.0.0
networkx==2.3
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603487797006/work
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.4.58
opencv-python-headless @ file:///home/yyw/yyf/projects/opencv_python_headless-4.5.4.60-cp36-cp36m-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
packaging==21.2
pandas==1.1.5
Pillow==6.0.0
progress==1.6
protobuf==3.7.1
pthflops==0.4.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2021.3
PyWavelets==1.0.3
PyYAML==6.0
pyzmq==18.0.1
qudida==0.0.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-build==0.12.0
scikit-image==0.15.0
scikit-learn==0.24.2
scipy==1.2.1
Shapely==1.8.0
six==1.12.0
sympy==1.9
tabulate==0.8.3
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
tensorpack==0.8.2
termcolor==1.1.0
terminaltables==3.1.0
thop==0.0.31.post2005241907
threadpoolctl==3.0.0
tifffile==2020.9.3
timm==0.4.12
torch==1.7.1
torch-tb-profiler==0.3.1
torchaudio==0.7.0a0+a853dff
torchsummary==1.5.1
torchvision==0.8.2
tqdm==4.31.1
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.26.7
Werkzeug==0.15.2
xmltodict==0.12.0
yapf==0.31.0
zipp==3.6.0
